Se agrego PathVariable

+++ src/main/java/com/Dialisis/DialisisPeritoneal/service/dto/PacienteInDto.java
@@ -5,15 +5,14 @@ import lombok.Data;
 import java.util.Date;
 
 @Data
-
 public class PacienteInDto extends UsuarioInDto {
     private long cedula;
     private Date fechaNacimiento;
     private int edad;
     private String eps;
     private int altura;
-    private int peso;
-    private int peso_seco;
+    private double peso;
+    private double peso_seco;
     private String direccion;
     private String ocupacion;
     private String tipo_sangre;
